+++ docs/conf.py
@@ -203,3 +203,7 @@ if eval(setup_cfg.get('edit_on_github')):
 # -- Options for the sphinx-argparse extension --------------------------------
 
 extensions += ['sphinxarg.ext']
+
+# -- Options for the sphinx-doctest extension ---------------------------------
+
+extensions += ['sphinx.ext.doctest']